% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling_triple_barrier.R
\name{add_vertical_barrier}
\alias{add_vertical_barrier}
\title{add_vertical_barrier}
\usage{
add_vertical_barrier(
  t_events,
  close_date,
  num_days = 1,
  num_hours = 0,
  num_minutes = 0,
  num_seconds = 0
)
}
\arguments{
\item{t_events}{Filtered events (e.g. feom CUSUM filter)}

\item{close_date}{Poxscit object of close vector or some other price vector}

\item{num_days}{Number of days into the future}

\item{num_hours}{Number of hours into the future}

\item{num_minutes}{Number of minutes into the future}

\item{num_seconds}{Number of seconds into the future}
}
\value{
data.table with 2 columns: t_events and vertical barrier
}
\description{
Compute vertical barriers
}
\examples{
data("spy")
cusum_events <- cusum_filter(subset(spy, select = c("index", "close")), 0.001)
add_vertical_barrier(cusum_events, spy$index, num_days = 1)

}
