% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling_triple_barrier.R
\name{apply_pt_sl_on_t1}
\alias{apply_pt_sl_on_t1}
\title{apply_pt_sl_on_t1}
\usage{
apply_pt_sl_on_t1(price, events, pt_sl = c(1, 1))
}
\arguments{
\item{price}{xts or data.table with index and price columns}

\item{events}{Indices that signify "events" (e.g. CUSUM events)}

\item{pt_sl}{Element 0, indicates the profit taking level; Element 1 is stop loss level}
}
\value{
data.table; timestamps of when first barrier was touched
}
\description{
Exponential moving average standard deviation
}
