% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatility.R
\name{daily_volatility}
\alias{daily_volatility}
\title{daily_volatility}
\usage{
daily_volatility(price, span = 50)
}
\arguments{
\item{price}{xts object with prices or data.table object with first column is of type POSIXct}

\item{span}{used to calculate alpha parameter in exponential moving average}
}
\value{
data.table with Datetime and Value columns
}
\description{
Advances in Financial Machine Learning, Snippet 3.1, page 44.
 Computes the daily volatility at intraday estimation points.
}
\examples{
data("spy")
daily_volatility(subset(spy, select = c("index", "close")))

}
