% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_gpd.R
\name{estimate_gpd}
\alias{estimate_gpd}
\title{Rolling GPD Help Function}
\usage{
estimate_gpd(x, hill_threshold = 0.02, suffix = "_right")
}
\arguments{
\item{x}{vector of (filtered) returns}

\item{hill_threshold}{look at ptsuite package}

\item{suffix}{suffix for column names}
}
\value{
Data.table with new features
}
\description{
Help function for calculating GPD features.
}
