% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_from_ohlcv.R
\name{features_from_ohlcv}
\alias{features_from_ohlcv}
\title{Genereta Features from OHLCV}
\usage{
features_from_ohlcv(
  ohlcv,
  window_sizes = c(5, 22),
  quantile_divergence_window = c(50, 100)
)
}
\arguments{
\item{ohlcv}{a data.table object with coluimns: symbol, datetime, open, hogh, low, close, volume}

\item{window_sizes}{Length of window for calculating rolling versions of the indicators.}
}
\value{
Data.table with new features
}
\description{
Function calculates basic features from OHLCV financial data
}
