% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_gas.R
\name{get_series_statistics}
\alias{get_series_statistics}
\title{Help function for Rolling GAS}
\usage{
get_series_statistics(df, colname_prefix = "var")
}
\arguments{
\item{df}{a data.table object with GAS results.}

\item{colname_prefix}{prefix for column names}
}
\value{
Data.table with GAS results
}
\description{
Help function for Rolling GAS
}
