% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling_abnormal_returns.R
\name{labeling_abnormal_returns}
\alias{labeling_abnormal_returns}
\title{Abnormal returns labeling}
\usage{
labeling_abnormal_returns(
  events,
  firm_returns,
  market_returns,
  event_horizon = 30,
  model_type = c("marketModel", "excessReturn", "None")
)
}
\arguments{
\item{events}{a data.frame of two columns with event dates (colname: “when”) and column names of the ‘response’ series from ‘firm.returns’ (colname “name”).}

\item{firm_returns}{zoo matrix of ‘outcome’ or ‘response’ series}

\item{market_returns}{Returns of SPY or some similiar asset.}

\item{event_horizon}{an ‘integer’ of length 1 that specifies a symmetric event window around the event time as specified in the index of “firm.returns”.}

\item{model_type}{Model type. See eventstudy function.}
}
\value{
Two data frames. First contains labels and abnormal returns and second contains whole returns path.
}
\description{
Define labels based on abnormal returns extracted from firm returns and events.
}
