% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling_around_zero.R
\name{labeling_around_zero}
\alias{labeling_around_zero}
\title{Laveling around zero}
\usage{
labeling_around_zero(x)
}
\arguments{
\item{x}{returns}
}
\value{
factor vector of labels
}
\description{
Make 3 labels: 1, 0, -1, where 0 are returns closest to zero
}
