% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{one_column_check}
\alias{one_column_check}
\title{one_column_check}
\usage{
one_column_check(x)
}
\arguments{
\item{x}{data.frame, xts or data.table object}
}
\value{
data.table with 1 column: Datetime
}
\description{
Help function to check arguments of x with one column.
}
\examples{
 data(spy)
 one_column_check(subset(spy, select = c("index")))
 \dontrun{one_column_check(subset(spy, select = c("close")))}

}
