% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_backcusum.R
\name{rolling_backcusum}
\alias{rolling_backcusum}
\title{Rolling BackCUSUM Tests}
\usage{
rolling_backcusum(
  prices_panel,
  row_index = 1:nrow(prices_panel),
  windows = c(200),
  workers = 4L
)
}
\arguments{
\item{prices_panel}{a data.table object with coluimns: symbol, datetime, open, hogh, low, close, volume}

\item{row_index}{row index for which to calculate radf values.}

\item{windows}{Length of window for calculating forecasts on rolling window.}

\item{workers}{Number of workers for parallel processing}
}
\value{
Data.table with new features
}
\description{
Function calculates backCUSUM tests on rolling window from backCUSUM pakcage.
}
