% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_catch22.R
\name{rolling_catch22}
\alias{rolling_catch22}
\title{Rolling catch 22}
\usage{
rolling_catch22(
  prices_panel,
  row_index = 1:nrow(prices_panel),
  windows = c(200),
  workers = 4L
)
}
\arguments{
\item{prices_panel}{a data.table object with coluimns: symbol, datetime, open, hogh, low, close, volume}

\item{windows}{Length of window for calculating forecasts on rolling window.}

\item{workers}{Number of workers for parallel processing}
}
\value{
Data.table with new features
}
\description{
Function calculates catch 22 features on rolling window.
}
