% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_exuber.R
\name{rolling_exuber}
\alias{rolling_exuber}
\title{Rolling Exuber}
\usage{
rolling_exuber(
  prices_panel,
  row_index = 1:nrow(prices_panel),
  windows = c(200),
  workers = 4L,
  exuber_lag = 0L
)
}
\arguments{
\item{prices_panel}{a data.table object with columns: symbol, datetime, close.}

\item{row_index}{row index for which to calculate radf values.}

\item{windows}{Length of window for calculating forecasts on rolling window.}

\item{workers}{Number of workers for parallel processing}

\item{exuber_lag}{exuber lag.}
}
\value{
Data.table with new features
}
\description{
Function calculates radf values from exuber package on rolling window.
}
