% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_gas.R
\name{rolling_gas}
\alias{rolling_gas}
\title{Rolling GAS}
\usage{
rolling_gas(
  prices_panel,
  row_index = 1:nrow(prices_panel),
  windows = c(200),
  workers = 4L,
  gas_dist = "sstd",
  gas_scaling = "Identity"
)
}
\arguments{
\item{prices_panel}{a data.table object with columns: symbol, datetime, close.}

\item{row_index}{row index for which to calculate radf values.}

\item{windows}{Length of window for calculating forecasts on rolling window.}

\item{workers}{Number of workers for parallel processing}

\item{gas_dist}{Dist parametere in UniGASSpec fucntionUniGASSpec.}

\item{gas_scaling}{Scaling parametere in UniGASSpec fucntionUniGASSpec.}
}
\value{
Data.table with new features
}
\description{
Function calculates GAS risk values from GAS package on rolling window.
}
