% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_gpd.R
\name{rolling_gpd}
\alias{rolling_gpd}
\title{Rolling GPD}
\usage{
rolling_gpd(
  prices_panel,
  row_index = 1:nrow(prices_panel),
  windows = c(200),
  workers = 4L,
  threshold = 0.03
)
}
\arguments{
\item{prices_panel}{a data.table object with coluimns: symbol, datetime, open, hogh, low, close, volume}

\item{row_index}{row index for which to calculate radf values.}

\item{windows}{Length of window for calculating forecasts on rolling window.}

\item{workers}{Number of workers for parallel processing}

\item{threshold}{threshold parameter for returns. Returns below/above this threshold will be input.}
}
\value{
Data.table with new features
}
\description{
Function calculates GPD risk measures on rolling window from ptsuite pakcage.
}
