% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_theft.R
\name{rolling_theft}
\alias{rolling_theft}
\title{Rolling catch 22}
\usage{
rolling_theft(
  prices_panel,
  row_index = 1:nrow(prices_panel),
  windows = c(200),
  workers = 8L,
  features_set = c("catch22", "feasts", "tsfeatures", "kats", "tsfresh", "tsfel")
)
}
\arguments{
\item{prices_panel}{a data.table object with coluimns: symbol, datetime, open, hogh, low, close, volume}

\item{row_index}{row index for which to calculate radf values.}

\item{windows}{Length of window for calculating forecasts on rolling window.}

\item{workers}{Number of workers for parallel processing}

\item{features_set}{Look at teft package}
}
\value{
Data.table with new features
}
\description{
Function calculates catch 22 features on rolling window.
}
