% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_univariate_forecasts.R
\name{rolling_univariate_forecasts}
\alias{rolling_univariate_forecasts}
\title{Auto Arima and Nnetar Forecasts}
\usage{
rolling_univariate_forecasts(
  prices_panel,
  row_index = 1:nrow(prices_panel),
  windows = c(200),
  workers = 4L,
  forecast_type = c("autoarima", "nnetar")
)
}
\arguments{
\item{prices_panel}{a data.table object with coluimns: symbol, datetime, open, hogh, low, close, volume}

\item{windows}{Length of window for calculating forecasts on rolling window.}

\item{workers}{Number of workers for parallel processing}
}
\value{
Data.table with new features
}
\description{
Function calculates forecastas based on auto.arima and nnetar functions from forecast package.
}
