% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{two_column_check}
\alias{two_column_check}
\title{two_column_check}
\usage{
two_column_check(x)
}
\arguments{
\item{x}{data.frame, xts or data.table object}
}
\value{
data.table with 2 columns: Datetime and Value
}
\description{
Help function to check arguments.
}
\examples{
 data(spy)
 two_column_check(subset(spy, select = c("index", "close")))
 \dontrun{two_column_check(subset(spy, select = c("index")))}
 \dontrun{two_column_check(subset(spy, select = c("close")))}
}
