% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{cusum_filter}
\alias{cusum_filter}
\title{cusum_filter}
\usage{
cusum_filter(price, threshold)
}
\arguments{
\item{price}{vector of prices}

\item{threshold}{scalar or numeric vector of same size as price.
When the change is larger than the threshold, the function captures it as an even}
}
\value{
vector of cusum events
}
\description{
Advances in Financial Machine Learning, Snippet 2.4, page 39.
 The Symmetric Dynamic/Fixed CUSUM Filter.
}
\examples{
data("spy")
cusum_events <- cusum_filter(spy$close, 0.001)
cusum_events <- spy$index[cusum_events]

}
