% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatility.R
\name{ewmsd}
\alias{ewmsd}
\title{ewmsd}
\usage{
ewmsd(y, alpha)
}
\arguments{
\item{y}{prices}

\item{span}{used to calculate alpha parameter in exponential moving average}
}
\value{
ewm standard devation
}
\description{
Exponential moving average standard devation
}
\examples{
data("spy")
daily_volatility(subset(spy, select = c("index", "close")))
}
